requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-filter==22.1
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.0.0
Markdown==3.4.1
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.11.0
